move methods to ColorUtils
