+++ mzmine-community/src/main/java/io/github/mzmine/modules/visualization/projectmetadata/color/ColorByMetadataTask.java
@@ -25,8 +25,6 @@
 
 package io.github.mzmine.modules.visualization.projectmetadata.color;
 
-import static io.github.mzmine.util.MathUtils.within;
-
 import io.github.mzmine.datamodel.RawDataFile;
 import io.github.mzmine.gui.MZmineGUI;
 import io.github.mzmine.javafx.concurrent.threading.FxThread;
@@ -40,6 +38,7 @@ import io.github.mzmine.modules.visualization.projectmetadata.table.columns.Meta
 import io.github.mzmine.parameters.ParameterSet;
 import io.github.mzmine.project.ProjectService;
 import io.github.mzmine.taskcontrol.AbstractRawDataFileTask;
+import io.github.mzmine.util.color.ColorUtils;
 import io.github.mzmine.util.color.SimpleColorPalette;
 import java.time.Instant;
 import java.util.Arrays;
@@ -54,11 +53,6 @@ import javafx.scene.paint.Color;
 import org.jetbrains.annotations.NotNull;
 
 public class ColorByMetadataTask extends AbstractRawDataFileTask {
-
-  // make sure brightness works for dark and light mode
-  public static final double minBrightness = 0.33;
-  public static final double maxBrightness = 0.75;
-  public static final double minSaturation = 0.35;
 
   private final List<RawDataFile> raws;
   private final Map<RawDataFile, Color> map = new HashMap<>();
@@ -89,20 +83,12 @@ public class ColorByMetadataTask extends AbstractRawDataFileTask {
     separateBlankQcs = parameters.getValue(ColorByMetadataParameters.separateBlankQcs);
     applySorting = parameters.getValue(ColorByMetadataParameters.applySorting);
     // as a percentage of the maximum
-    brightnessPercentRange = maxBrightnessWidth() * parameters.getValue(
+    brightnessPercentRange = ColorUtils.maxBrightnessWidth() * parameters.getValue(
         ColorByMetadataParameters.brightnessPercentRange);
 
     colors = ConfigService.getDefaultColorPalette().clone(true);
   }
 
-
-  public static double maxBrightnessWidth() {
-    return maxBrightness - minBrightness;
-  }
-
-  public static double centerBrightness() {
-    return (maxBrightness + minBrightness) / 2d;
-  }
 
   @Override
   protected void process() {
@@ -198,83 +184,15 @@ public class ColorByMetadataTask extends AbstractRawDataFileTask {
       usedColors.add(base); // mark as used
       return;
     }
-    if (base.getSaturation() == 0) {
-      colorFadeGray(raws);
-      return;
-    }
 
-    usedColors.add(base); // mark as used
+    usedColors.add(base);
+    var colors = ColorUtils.colorFadeLighter(raws.size(), base, bRange);
 
-    double h = base.getHue();
-
-    // start saturation --> 1  - tend to higher saturations rather than lower
-    double sRange = bRange * 1.5;
-    double maxS = within(base.getSaturation() + 0.1 + sRange, minSaturation, 1);
-    double minS = maxS - sRange;
-    if (minS < minSaturation) {
-      minS = minSaturation;
-      maxS = within(minS + sRange, minSaturation, 1);
-    }
-
-    // start in center brightness to allow both light and dark mode
-    // prefer lighter colors
-    double maxB = within(base.getBrightness() + bRange * 0.6, minBrightness, maxBrightness);
-    double minB = maxB - bRange;
-    if (minB < minBrightness) {
-      minB = minBrightness;
-      maxB = within(minB + bRange, minBrightness, maxBrightness);
-    }
-
-    // option to start always in the center - but this makes colors quite dull and they only depend on hue
-//    double minB = within(centerBrightness() - bRange / 2d, minBrightness, maxBrightness);
-//    double maxB = within(minB + bRange, minBrightness, maxBrightness);
-
-    // first half only increases brightness with max saturation
-    double halfN = Math.max(Math.floor(raws.size() / 2d), 1);
-    double stepB = (maxB - minB) / (halfN);
-    double b = minB;
-    int step = 0;
-    for (; step < halfN; step++) {
-      map.put(raws.get(step), Color.hsb(h, maxS, b));
-      b = within(b + stepB, minB, maxB);
-    }
-
-    // reduce saturation
-    double stepS = (maxS - minS) / Math.max(raws.size() - halfN - 1, 1);
-    double s = maxS;
-    for (; step < raws.size(); step++) {
-      map.put(raws.get(step), Color.hsb(h, s, maxB));
-      s = within(s - stepS, minS, maxS); // need to modify after adding the last brightness step
+    for (int i = 0; i < raws.size(); i++) {
+      map.put(raws.get(i), colors.get(i)); // actual recoloring
     }
   }
 
-  private void colorFadeGray(final List<RawDataFile> raws) {
-    colorFadeGray(raws, brightnessPercentRange);
-  }
-
-  /**
-   * Color around 50% gray +- bRange/2
-   *
-   * @param bRange brightness width
-   */
-  private void colorFadeGray(final List<RawDataFile> raws, double bRange) {
-    if (raws.isEmpty()) {
-      return;
-    }
-    if (raws.size() == 1) {
-      map.put(raws.getFirst(), Color.hsb(0, 0, centerBrightness()));
-      return;
-    }
-    int n = Math.max(raws.size() - 1, 1);
-    bRange = Math.min(bRange, maxBrightnessWidth());
-    double maxB = within(centerBrightness() + bRange / 2d, minBrightness, maxBrightness);
-
-    int step = 0;
-    for (RawDataFile raw : raws) {
-      map.put(raw, Color.hsb(0, 0, maxB - bRange * (step / (double) n)));
-      step++;
-    }
-  }
 
   @Override
   protected void addAppliedMethod() {
