+++ mzmine-community/src/main/java/io/github/mzmine/util/color/ColorUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2004-2024 The MZmine Development Team
+ * Copyright (c) 2004-2024 The mzmine Development Team
  *
  * Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
@@ -25,12 +25,20 @@
 
 package io.github.mzmine.util.color;
 
+import static io.github.mzmine.util.MathUtils.within;
+
 import io.github.mzmine.javafx.util.FxColorUtil;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.logging.Logger;
 import javafx.scene.paint.Color;
 
 public class ColorUtils {
 
+  // make sure brightness works for dark and light mode
+  public static final double minBrightness = 0.33;
+  public static final double maxBrightness = 0.75;
+  public static final double minSaturation = 0.35;
   private static final Logger logger = Logger.getLogger(ColorUtils.class.getName());
 
   /**
@@ -112,5 +120,100 @@ public class ColorUtils {
 
   public static boolean isDark(Color clr) {
     return getColorDifference(clr, Color.BLACK) < 250;
+  }
+
+  public static double maxBrightnessWidth() {
+    return maxBrightness - minBrightness;
+  }
+
+  public static double centerBrightness() {
+    return (maxBrightness + minBrightness) / 2d;
+  }
+
+  /**
+   * Color around the base color. Brightness and saturation are scaled
+   *
+   * @param bRange brightness width
+   */
+  public static List<Color> colorFadeLighter(final int steps, final Color base,
+      final double bRange) {
+    if (steps == 0) {
+      return List.of();
+    }
+    if (steps == 1) {
+      return List.of(base);
+    }
+
+    if (base.getSaturation() == 0) {
+      return colorFadeGray(steps, bRange);
+    }
+
+    List<Color> resultColors = new ArrayList<>(steps);
+
+    double h = base.getHue();
+
+    // start saturation --> 1  - tend to higher saturations rather than lower
+    double sRange = bRange * 1.5;
+    double maxS = within(base.getSaturation() + 0.1 + sRange, minSaturation, 1);
+    double minS = maxS - sRange;
+    if (minS < minSaturation) {
+      minS = minSaturation;
+      maxS = within(minS + sRange, minSaturation, 1);
+    }
+
+    // start in center brightness to allow both light and dark mode
+    // prefer lighter colors
+    double maxB = within(base.getBrightness() + bRange * 0.6, minBrightness, maxBrightness);
+    double minB = maxB - bRange;
+    if (minB < minBrightness) {
+      minB = minBrightness;
+      maxB = within(minB + bRange, minBrightness, maxBrightness);
+    }
+
+    // option to start always in the center - but this makes colors quite dull and they only depend on hue
+//    double minB = within(centerBrightness() - bRange / 2d, minBrightness, maxBrightness);
+//    double maxB = within(minB + bRange, minBrightness, maxBrightness);
+
+    // first half only increases brightness with max saturation
+    double halfN = Math.max(Math.floor(steps / 2d), 1);
+    double stepB = (maxB - minB) / (halfN);
+    double b = minB;
+    int step = 0;
+    for (; step < halfN; step++) {
+      resultColors.add(Color.hsb(h, maxS, b));
+      b = within(b + stepB, minB, maxB);
+    }
+
+    // reduce saturation
+    double stepS = (maxS - minS) / Math.max(steps - halfN - 1, 1);
+    double s = maxS;
+    for (; step < steps; step++) {
+      resultColors.add(Color.hsb(h, s, maxB));
+      s = within(s - stepS, minS, maxS); // need to modify after adding the last brightness step
+    }
+    return resultColors;
+  }
+
+  /**
+   * Color around 50% gray +- bRange/2
+   *
+   * @param bRange brightness width
+   */
+  public static List<Color> colorFadeGray(final int steps, double bRange) {
+    if (steps == 0) {
+      return List.of();
+    }
+    if (steps == 1) {
+      return List.of(Color.hsb(0, 0, centerBrightness()));
+    }
+    int n = Math.max(steps - 1, 1);
+    bRange = Math.min(bRange, maxBrightnessWidth());
+    double maxB = within(centerBrightness() + bRange / 2d, minBrightness, maxBrightness);
+
+    List<Color> resultColors = new ArrayList<>(steps);
+    for (int step = 0; step < steps; step++) {
+      resultColors.add(Color.hsb(0, 0, maxB - bRange * (step / (double) n)));
+    }
+    return resultColors;
   }
 }
